Refactor: styock details screen with argumens
