+++ lib/core/navigators/router.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:stock/core/constants/routes.dart';
 import 'package:stock/ui/features/auth/presentation/pages/login_screen.dart';
+import 'package:stock/ui/features/home/domain/entities/stock_entity.dart';
 import 'package:stock/ui/features/home/presentation/pages/market_details.dart';
-import 'package:stock/ui/features/home/presentation/pages/tab_screen.dart';
+import 'package:stock/ui/features/home/presentation/pages/tab_screen_screen.dart';
 import 'package:stock/ui/features/introduction/splash_screen.dart';
 
 /// Generate routes for navigation
@@ -26,7 +27,9 @@ Route<dynamic> generateRoute(RouteSettings settings) {
     case Routes.marketDetailView:
       return _getPageRoute(
         routeName: settings.name,
-        viewToShow: const MarketDetailScreen(),
+        viewToShow: MarketDetailScreen(
+          params: settings.arguments as StockEntity,
+        ),
       );
     default:
       return MaterialPageRoute(
