+++ lib/ui/features/home/presentation/widgets/single_market_widget.dart
@@ -18,6 +18,7 @@ class SingleMarketWidget extends StatelessWidget {
       onTap: () => Navigator.pushNamed(
         context,
         Routes.marketDetailView,
+        arguments: item,
       ),
       child: Container(
         margin: const EdgeInsets.only(bottom: 15),
